+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Add "Remove argument from predicate/non-terminal" refactoring support
 * Add "Extract to new Logtalk entity" refactoring support
 * Add "Extract to new Logtalk file" refactoring support
+* Add "Replace with include/1 directive" refactoring support
 * Add "Replace include/1 directive with file contents" refactoring support
 * Add experimental entity, predicate, and non-terminal rename support
 * Add "Logtalk: Compute Project Metrics" command